miner/requirements/requirements_cu124.txt
biopython==1.85
bittensor==9.0.1
datasets==3.3.2
fair-esm==2.0.0
huggingface-hub==0.29.1
lifelines==0.30.0
pandas==2.2.3
python-dotenv==1.0.1
rdkit==2024.9.4
requests==2.32.3
substrate-interface==1.7.11
scikit-learn==1.6.1
# GPU optimization libraries
cupy-cuda12x>=12.0.0
nvtx>=0.2.5
triton>=2.0.0
numba>=0.58.0
deepspeed>=0.11.0
torch-memory-efficient-attention>=0.1.0